+++ src/main/java/kr/tennispark/activity/admin/application/impl/ActivityImageAdminService.java
@@ -3,7 +3,7 @@ package kr.tennispark.activity.admin.application.impl;
 
 import kr.tennispark.activity.admin.infrastructure.repository.ActivityImageRepository;
 import kr.tennispark.activity.common.domain.ActivityImage;
-import kr.tennispark.qr.application.S3UploadService;
+import kr.tennispark.common.application.S3UploadService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
