+++ src/main/java/kr/tennispark/common/application/S3UploadService.java
@@ -1,4 +1,4 @@
-package kr.tennispark.qr.application;
+package kr.tennispark.common.application;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
